+++ package.json
@@ -4,7 +4,7 @@
     "homepage": ".",
     "private": true,
     "dependencies": {
-        "@gridsuite/commons-ui": "^0.8.0",
+        "@gridsuite/commons-ui": "^0.9.0",
         "@material-ui/core": "^4.9.5",
         "@material-ui/icons": "^4.5.1",
         "@material-ui/lab": "^4.0.0-alpha.42",
@@ -22,9 +22,7 @@
         "react-window": "^1.8.5",
         "reconnecting-websocket": "^4.4.0",
         "redux": "^4.0.5",
-        "typeface-roboto": "0.0.75",
-        "react-virtualized": "^9.21.2",
-        "memoize-one": "^5.1.1"
+        "typeface-roboto": "0.0.75"
     },
     "scripts": {
         "start": "react-scripts start",
